Revert revision 38013 and revision 38011
They are not working as expected

Review URL: https://codereview.chromium.org//368383002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@38017 260f80e4-7a28-3924-810f-c04153c831b5

+++ pkg/polymer/lib/src/build/runner.dart
@@ -216,21 +216,7 @@ Future _emitAllFiles(Barback barback, BarbackOptions options) {
   return barback.getAllAssets().then((assets) {
     // Delete existing output folder before we generate anything
     var dir = new Directory(options.outDir);
-    // We get very long filenames which can cause us to not be able to delete
-    // using the normal dart:io deleteSync(recursive: true) on windows
-    if (dir.existsSync()) {
-      if (Platform.operatingSystem == 'windows' ) {
-        var result = Process.runSync('rmdir', ['/q', '/s', options.outDir],
-                                     runInShell: true);
-        if (result.exitCode != 0) {
-          throw "Could not delete $dir, output was: \n"
-                 "stdout: ${result.stdout} \n"
-                 "stderr: ${result.stderr}";
-        }
-      } else {
-        dir.deleteSync(recursive: true);
-      }
-    }
+    if (dir.existsSync()) dir.deleteSync(recursive: true);
     return _emitPackagesDir(options)
       .then((_) => _emitTransformedFiles(assets, options))
       .then((_) => _addPackagesSymlinks(assets, options))
